Moved `Bookmark` into public API and then provided a string conversion method to convert string bookmarks to `Bookmark`
