+++ driver/src/main/java/org/neo4j/driver/Bookmark.java
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.neo4j.driver.internal;
+package org.neo4j.driver;
 
 import java.io.Serializable;
