+++ driver/src/main/java/org/neo4j/driver/Session.java
@@ -21,7 +21,6 @@ package org.neo4j.driver;
 import java.util.Map;
 
 import org.neo4j.driver.async.AsyncSession;
-import org.neo4j.driver.internal.Bookmark;
 import org.neo4j.driver.util.Resource;
 
 /**
