+++ driver/src/main/java/org/neo4j/driver/internal/InternalBookmark.java
@@ -25,6 +25,8 @@ import java.util.HashSet;
 import java.util.Objects;
 import java.util.Set;
 
+import org.neo4j.driver.Bookmark;
+
 import static java.util.Objects.requireNonNull;
 
 public final class InternalBookmark implements Bookmark
@@ -97,7 +99,7 @@ public final class InternalBookmark implements Bookmark
     {
         if ( !(bookmark instanceof InternalBookmark) )
         {
-            throw new IllegalArgumentException( String.format( "Received bookmark '%s' is not generated by driver sessions.", bookmark ) );
+            throw new IllegalArgumentException( String.format( "Received bookmark '%s' is not generated by a driver session.", bookmark ) );
         }
     }
