+++ driver/src/main/java/org/neo4j/driver/internal/InternalSession.java
@@ -21,6 +21,7 @@ package org.neo4j.driver.internal;
 import java.util.Map;
 
 import org.neo4j.driver.AccessMode;
+import org.neo4j.driver.Bookmark;
 import org.neo4j.driver.Session;
 import org.neo4j.driver.Statement;
 import org.neo4j.driver.StatementResult;
