+++ driver/src/test/java/org/neo4j/driver/SessionConfigTest.java
@@ -29,8 +29,6 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Stream;
 
-import org.neo4j.driver.internal.Bookmark;
-
 import static java.util.Collections.emptyList;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.junit.Assert.assertEquals;
