+++ driver/src/test/java/org/neo4j/driver/stress/AbstractContext.java
@@ -21,7 +21,7 @@ package org.neo4j.driver.stress;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 
-import org.neo4j.driver.internal.Bookmark;
+import org.neo4j.driver.Bookmark;
 import org.neo4j.driver.summary.ResultSummary;
 
 public abstract class AbstractContext
