+++ driver/src/test/java/org/neo4j/driver/util/BookmarkUtilTest.java
@@ -1,0 +1,51 @@
+/*
+ * Copyright (c) 2002-2019 "Neo4j,"
+ * Neo4j Sweden AB [http://neo4j.com]
+ *
+ * This file is part of Neo4j.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.neo4j.driver.util;
+
+import org.junit.jupiter.api.Test;
+
+import java.util.Arrays;
+
+import org.neo4j.driver.Bookmark;
+import org.neo4j.driver.internal.InternalBookmark;
+
+import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.CoreMatchers.hasItems;
+import static org.hamcrest.CoreMatchers.instanceOf;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.neo4j.driver.internal.util.BookmarkUtil.*;
+
+class BookmarkUtilTest
+{
+    @Test
+    void shouldParseSingleBookmarkString() throws Throwable
+    {
+        Bookmark bookmark = BookmarkUtil.parse( "a bookmark, another bookmark" );
+
+        assertBookmarkContainsSingleValue( bookmark, equalTo( "a bookmark, another bookmark" ) );
+    }
+
+    @Test
+    void shouldParseBookmarkStringCollections() throws Throwable
+    {
+        Bookmark bookmark = BookmarkUtil.parse( Arrays.asList( "a bookmark", "another bookmark" ) );
+
+        assertBookmarkContainsValues( bookmark, hasItems( "a bookmark", "another bookmark" ) );
+    }
+}
